requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
pyyaml>=6.0

# Deep learning frameworks
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.9.0

# API and web framework
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.10.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymongo>=4.2.0
redis>=4.3.0

# Optimization libraries
optuna>=3.0.0
hyperopt>=0.2.7
scipy>=1.9.0

# Utilities
python-multipart>=0.0.5
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Development and testing
pytest>=7.1.0
pytest-asyncio>=0.19.0
black>=22.6.0
flake8>=5.0.0
mypy>=0.971

matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Monitoring and health checks
psutil>=5.8.0
aiohttp>=3.8.0
GPUtil>=1.4.0